% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add}
\alias{add}
\title{Add Two Numbers}
\usage{
add(i, j)
}
\arguments{
\item{i}{A numeric value representing the first number.}

\item{j}{A numeric value representing the second number.}
}
\value{
A numeric value equal to the sum of \code{i} and \code{j}.
}
\description{
This function takes two numeric inputs, \code{i} and \code{j}, and returns their sum.
}
