\name{calc-package}
\alias{calc-package}
\alias{calc}
\docType{package}
\title{
\packageTitle{calc}
}
\description{
\packageDescription{calc}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{calc}
\packageIndices{calc}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{calc}

Maintainer: \packageMaintainer{calc}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the ~~
~~ R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
