% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dif.R
\name{dif}
\alias{dif}
\title{Subtract Two Numbers}
\usage{
dif(i, j)
}
\arguments{
\item{i}{A numeric value representing the first number.}

\item{j}{A numeric value representing the second number to subtract from \code{i}.}
}
\value{
A numeric value equal to \code{i - j}.
}
\description{
This function subtracts the second numeric input \code{j} from the first numeric input \code{i}.
}
