% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div.R
\name{div}
\alias{div}
\title{Divide Two Numbers}
\usage{
div(i, j)
}
\arguments{
\item{i}{A numeric value representing the numerator.}

\item{j}{A numeric value representing the denominator.}
}
\value{
A numeric value equal to \code{i / j}.
}
\description{
This function divides the first numeric input \code{i} by the second numeric input \code{j}.
If \code{j} is zero, the function will return \code{Inf} or \code{NaN}.
}
