% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mul.R
\name{mul}
\alias{mul}
\title{Multiply Two Numbers}
\usage{
mul(i, j)
}
\arguments{
\item{i}{A numeric value representing the first number.}

\item{j}{A numeric value representing the second number.}
}
\value{
A numeric value equal to \code{i * j}.
}
\description{
This function multiplies two numeric inputs, \code{i} and \code{j}.
}
